feat: restructuring the project
