+++ src/dsAuth.ts
@@ -1,9 +1,12 @@
-import { Client } from 'discord.js';
-import ready from '@events/ready';
+import { CustomClient } from './interfaces/CustomClient';
+import { Collection } from 'discord.js';
+import { calendar_v3 } from 'googleapis';
 
-export default async function dsAuth() {
+export default async function dsAuth(calendar: calendar_v3.Calendar) {
     // Create a new client instance
-    const client: Client = new Client({ intents: 1 });
+    const client: CustomClient = new CustomClient({ intents: 1 }, calendar);
+
+
 
     await client.login(process.env.DISCORD_TOKEN).then(() => { console.log('logged'); });
